requirements.txt
pypandoc==1.13
pypdf==4.0.1
reportlab==4.1.0